fastCollinear fixed
you should not re-create LineSegment[] to be returned in each method of requesting this array, since the original result stack will be eliminated after computation. Also, create a dummy array when returning the LineSegment[] finalResult in order to protect the original data from being mutated.

When copying elements in stack by popping, the limit of the loop should rely on a fixed value instead of the dynamic SIZE() of the stack!