solver beta 6.16
In this beta, each time the min element is removed from the PQ, I will make sure that it belongs to the parent group of current search depth. This means, when algorithms selects specific children to add its neighbors to the PQ, it can only choose one next step from these neighbors.